GenshinAI/requirements.txt
# Core dependencies
python-dotenv>=0.19.0
httpx>=0.24.0

# Discord bot
discord.py>=2.0.0

# Web scraping and API
beautifulsoup4>=4.10.0
aiohttp>=3.8.1

# Data handling
json5>=0.9.10

# Development tools
black==23.7.0
flake8==6.1.0
pytest>=7.0.0
pre-commit>=3.3.2
ruff>=0.0.285
pydocstyle>=6.3.0
pyupgrade>=3.10.1
